Fix the numpy_typemaps
ToHostMatrix has been broken for a while because of the DOUBLE typemap
that was in there.  Stick to single precision, since we're doing log
domain math anyhow.

[Tracking bug 23893]

[git-p4: depot-paths = "//depot/software/common/ConsensusCore/": change = 129463]

+++ src/C++/Quiver/Diploid.hpp
@@ -58,6 +58,6 @@ namespace ConsensusCore {
     //  (siteScores)
     // for Python by SWIG.
     DiploidSite*
-    IsSiteHeterozygous(const double *siteScores, int dim1, int dim2,
-                       double logPriorRatio);
+    IsSiteHeterozygous(const float *siteScores, int dim1, int dim2,
+                       float logPriorRatio);
 }
